[ADD] 코드 유효시간에 따라 분기처리(#34)
